Add new featute: extract cludiness from database * add script * add configs * add config reader script * update setup.py * update .gitignore
