+++ .gitignore
@@ -6,6 +6,9 @@ __pycache__/
 # C extensions
 *.so
 
+# private config file
+src/mordor/conf/*.private
+
 # Distribution / packaging
 .Python
 build/
